requirements.txt
streamlit>=1.30.0
groq>=0.4.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pytesseract>=0.3.10
pdf2image>=1.16.0
